+++ src/ngx_http_accounting_module.c
@@ -6,7 +6,7 @@
 
 #include <ngx_http.h>
 
-#include "ngxta.h"
+#include "ngx_traffic_accounting.h"
 #include "ngx_http_accounting_module.h"
 #include "ngx_http_accounting_worker_process.h"
